wait-process: omit confusing cast
* lib/wait-process.c (wait_subprocess): Omit confusing cast.

+++ lib/wait-process.c
@@ -383,7 +383,7 @@ wait_subprocess (pid_t child, const char *progname,
       if (exit_on_error || (!null_stderr && termsigp == NULL))
         error (exit_on_error ? EXIT_FAILURE : 0, 0,
                _("%s subprocess got fatal signal %d"),
-               progname, (int) WTERMSIG (status));
+               progname, WTERMSIG (status));
       return 127;
     }
   if (!WIFEXITED (status))
